Move stream impl
