+++ python/core/src/ffi/from_python/ffi_stream.rs
