+++ python/core/src/ffi/mod.rs
@@ -1,5 +1,4 @@
 //! Arrow FFI via the C Data Interface and the Arrow PyCapsule Interface.
 
 pub mod from_python;
-pub mod stream_chunked;
 pub mod to_python;
